requirements.txt
Django==1.7.8
django-bootstrap3==6.1.0
django-debug-toolbar==1.3.2
django-djconfig==0.3.2
django-haystack==2.3.1
django-infinite-scroll-pagination==0.1.3
django-spirit==0.2.1
djangorestframework==3.1.3
mistune==0.3.1
Pillow==2.6.2
pytz==2015.4
sqlparse==0.1.15
Whoosh==2.5.7
